Elminate `undef` warning.
It's something like "Use of uninitialized value in list assignment," but what
it's really complaining about is the use of an `undef` as a hash key. Seems
that `XML::LibXML::Namespace->localname` can return `undef`. So fall back on
the empty string. I didn't run into it in the `sprintf` line, but since using
an `undef` there also triggers a warning, it seemed prudent to fall back on
the empty string there, too.